Refactored out v3 endpoint.
