+++ src/applications/check-in/api/index.js
@@ -1,9 +1,7 @@
 import { v2 } from './versions/v2';
-import { v3 } from './versions/v3';
 
 const api = {
   v2,
-  v3,
 };
 
 export { api };
