+++ src/applications/check-in/api/local-mock-api/e2e/ApiInitializer.js
@@ -125,8 +125,8 @@ class ApiInitializer {
   };
 
   initializeSessionPost = {
-    withSuccess: (extraValidation, version = 'v2') => {
-      cy.intercept('POST', `/check_in/${version}/sessions`, req => {
+    withSuccess: extraValidation => {
+      cy.intercept('POST', `/check_in/v2/sessions`, req => {
         if (extraValidation) {
           extraValidation(req);
         }
