--- src/applications/check-in/api/versions/v3.js
@@ -1,36 +1,0 @@
-import { apiRequest } from 'platform/utilities/api';
-import environment from 'platform/utilities/environment';
-import { makeApiCallWithSentry } from '../utils';
-
-const v3 = {
-  postSession: async ({ lastName, dob, token, checkInType = '' }) => {
-    const url = '/check_in/v3/sessions/';
-    const headers = { 'Content-Type': 'application/json' };
-    const data = {
-      session: {
-        uuid: token,
-        dob,
-        lastName: lastName.trim(),
-        checkInType,
-      },
-    };
-    const body = JSON.stringify(data);
-    const settings = {
-      headers,
-      body,
-      method: 'POST',
-      mode: 'cors',
-    };
-
-    const json = await makeApiCallWithSentry(
-      apiRequest(`${environment.API_URL}${url}`, settings),
-      'validating-user',
-      token,
-    );
-    return {
-      ...json,
-    };
-  },
-};
-
-export { v3 };
