+++ src/applications/check-in/day-of/tests/e2e/dob-validation/dob.validation.cypress.spec.js
@@ -18,7 +18,7 @@ describe('Check In Experience ', () => {
     initializeCheckInDataGet.withSuccess({
       numberOfCheckInAbledAppointments: 1,
     });
-    initializeSessionPost.withSuccess(false, 'v3');
+    initializeSessionPost.withSuccess();
 
     initializeCheckInDataPost.withSuccess();
   });
